Bug fixes in preserving elements properties..
Note that there the speed gains in in the last commit for
`Do[F[a,a,a,a,a,a,a,a,a,a,a]` are almost completely gone.
(Not quite, so we still are better than 3 or so commits ago but worse
than 4.0.0 on this.

It might be cause by slowdown in eval_elements() and eval_range()
needing element properties to be recomputed, and some of this might be
unnecessary. I've marked these with FIXME.